requirements.txt
# ── Core LiveKit Agents framework ─────────────────────────────────────────────
livekit-agents[deepgram,openai,elevenlabs,silero,anthropic]>=0.12.0

# ── LiveKit Server SDK (token generation) ─────────────────────────────────────
livekit-api>=0.7.0

# ── Config & environment ──────────────────────────────────────────────────────
pydantic-settings>=2.3.0
python-dotenv>=1.0.0

# ── API server (token endpoint for portfolio integration) ─────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# ── Email ─────────────────────────────────────────────────────────────────────
resend>=0.8.0

# ── GitHub ────────────────────────────────────────────────────────────────────
PyGithub>=2.3.0

# ── Google Calendar & Meet ────────────────────────────────────────────────────
google-auth>=2.30.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.130.0

# ── Utilities ─────────────────────────────────────────────────────────────────
httpx>=0.27.0
aiofiles>=23.2.1
python-dateutil>=2.9.0
pytz>=2024.1